Change field type from integer to double

+++ database/migrations/2021_02_24_200035_create_things_table.php
@@ -18,7 +18,7 @@ class CreateThingsTable extends Migration
             $table->timestamps();
             $table->date('date');
             $table->string('name');
-            $table->integer('sum');
+            $table->double('sum');
         });
     }
